Release v0.6.12
Reconstructed from the shipped release archive. No version control
existed for this release; the archive is the only surviving record.

X-Provenance: e107v0.612.tar.gz
X-Provenance-SHA256: 28c60cdab1d61c62fe7a119ed5e1705f839530d09a7457cf6fb34209d2199489
X-Provenance-Files: 948
X-Author-Basis: source header in e107v0.612.tar.gz
X-Date-Basis: author=max member mtime; committer=SourceForge publication
X-Release-Kind: release
X-SourceForge-Folder: e107/e107 v0.612

+++ e107_admin/filemanager.php
@@ -22,6 +22,7 @@ $path = str_replace("../", "", (e_QUERY ? e_QUERY : e_FILE));
 if(!$path){ $path =  str_replace("../", "", e_FILE); }
 
 if(IsSet($_POST['deletefile'])){
+	if(!$_POST['ac'] == md5(ADMINPWCHANGE)){ exit; }
 	$destination_file = str_replace($ADMIN_DIRECTORY, "", substr($_SERVER['PATH_TRANSLATED'], 0, strrpos($_SERVER['PATH_TRANSLATED'], "/"))."/".$_POST['deleteconfirm']);
 	if(@unlink($destination_file)){
 		$message = "Deleted '".$destination_file."' successfully.";
@@ -31,6 +32,7 @@ if(IsSet($_POST['deletefile'])){
 }
 
 if(IsSet($_POST['upload'])){
+	if(!$_POST['ac'] == md5(ADMINPWCHANGE)){ exit; }
 	$files = $_FILES['userfile'];
 	foreach($files['name'] as $key => $name){
 		if($files['size'][$key]){
@@ -223,6 +225,7 @@ while($files[$c]){
 }
 
 $text .= "
+<input type='hidden' name='ac' value='".md5(ADMINPWCHANGE)."' />
 </form>
 </table>
 </div>";
